-clickAttribution (superceded by kifiHit)
